Tiny fixes in model and arguments

+++ train.py
@@ -90,7 +90,7 @@ def train(config, generator, discriminator, kp_detector, checkpoint, log_dir, da
                 losses = {key: value.mean().detach().data.cpu().numpy() for key, value in losses_generator.items()}
                 logger.log_iter(losses=losses)
 
-                step = i + epoch * len(dataset) / dataloader.batch_size
+                step = i + int(epoch * len(dataset) / dataloader.batch_size)
                 if step % 20 == 0:
                     print_fun(f'Epoch {epoch + 1}, global step {step}: {", ".join([f"{k}={v}" for k, v in losses.items()])}')
